fix: update Makefile to trigger rebuild on changes and improve solution.cpp type safety

+++ leetcode/3550/solution.cpp
@@ -11,8 +11,8 @@ public:
                 sum += number % 10;
                 number /= 10;
             }
-            if(sum == i) {
-                return i;
+            if(sum == static_cast<int>(i)) {
+                return static_cast<int>(i);
             }
         }
